services/admin-api/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# MongoDB async driver
motor==3.3.2
pymongo==4.6.1

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.2.1

# Environment and utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# HTTP client for API testing
httpx==0.26.0

# SSE for real-time log streaming
sse-starlette==1.8.2
